+++ src/main/java/subway/application/dto/SectionParam.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.application.dto;
 
 public class SectionParam {
