+++ src/main/java/subway/ui/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.List;
