+++ src/main/java/subway/ui/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto;
 
 import subway.domain.Station;
