+++ src/test/java/subway/application/SectionServiceTest.java
@@ -24,8 +24,8 @@ import subway.dao.StationDao;
 import subway.domain.Line;
 import subway.domain.Section;
 import subway.domain.Station;
-import subway.dto.SectionRequest;
-import subway.dto.SectionResponse;
+import subway.ui.dto.SectionRequest;
+import subway.ui.dto.SectionResponse;
 import subway.exception.IllegalLineException;
 import subway.exception.IllegalSectionException;
