+++ src/test/java/subway/application/StationServiceTest.java
@@ -20,8 +20,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import subway.dao.StationDao;
 import subway.domain.Station;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
+import subway.ui.dto.StationRequest;
+import subway.ui.dto.StationResponse;
 import subway.exception.IllegalStationsException;
 
 @DisplayName("역 서비스 테스트")
